requirements.txt
python-dotenv
pydantic-ai
ruff